requirements_test.txt
atomicwrites==1.3.0
attrs==18.2.0
certifi==2019.3.9
chardet==3.0.4
commonmark==0.8.1
dataclasses==0.6
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.0
django-preferences==1.0.0
django-rest-auth==0.9.3
django-stubs==0.9.0
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
more-itertools==5.0.0
mypy==0.670
mypy-extensions==0.4.1
pluggy==0.8.1
psycopg2==2.7.7
py==1.7.0
pytest==4.2.0
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
whitenoise==4.1.2
